Add possibility to run locally as simulation

+++ app/controller/CameraController.py
@@ -5,24 +5,33 @@ from time import sleep
 
 from app.constants import CAMERA_GPIO
 from app.controller import BaseController
+from app.utils import Environment
 
 RETRIES = 5
 
 
 class CameraController(BaseController):
+    env = Environment()
+
     def __init__(self):
         print(f"Connecting the camera on channel {CAMERA_GPIO}")
-        camera = cv2.VideoCapture(CAMERA_GPIO)
-        super().__init__(camera, CAMERA_GPIO)
-        self.connector.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
-        self.connector.camera.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
+        if self.env.is_live:
+            camera = cv2.VideoCapture(CAMERA_GPIO)
+            super().__init__(camera, CAMERA_GPIO)
+            self.connector.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
+            self.connector.camera.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
+        else:
+            super().__init__(None, CAMERA_GPIO)
         self.frame = None
         self.read_frame()
 
     def read_frame(self):
         retry_count = 1
         while True:
-            ret, self.frame = self.connector.read()
+            if self.env.is_live:
+                ret, self.frame = self.connector.read()
+            else:
+                ret, self.frame = True, None
             retry_count += 1
             if ret:
                 break
@@ -37,15 +46,17 @@ class CameraController(BaseController):
         return self.frame
 
     def display_current_frame(self):
-        # Display the frame using Matplotlib
-        cv2.imshow("Camera feed", self.current_frame)
+        if self.env.is_live:
+            # Display the frame using Matplotlib
+            cv2.imshow("Camera feed", self.current_frame)
 
     @abstractmethod
     def draw_circle(self, image, x, y, radius, rgb, z):
-        cv2.circle(image, (x, y), radius, rgb, z)
+        if self.env.is_live:
+            cv2.circle(image, (x, y), radius, rgb, z)
 
     def stop(self):
         print(f"Disconnecting the camera on channel {self.channel}")
-        self.connector.release()
-        cv2.destroyAllWindows()
-
+        if self.env.is_live:
+            self.connector.release()
+            cv2.destroyAllWindows()
